# Initial tag query performance improvements