@DanCech 2018-02-09
Very interesting, how does it impact memory usage for the map?