@shanson7 2018-02-09
The strings all reference the strings that each `def` already has. So, the keys themselves are just the size of a golang string (pointer plus int) so they are smaller than the existing keys.