@replay 2018-02-12
Interesting. Actually at the time of writing that, as you already mentioned, my idea was that the next step would be to change everything that's still using strings into numeric types and then we won't need to convert between strings and numeric types anymore. I did not expect this temporary solution of converting forth and back is going to be that expensive, so I think your changes make sense that way.